funcionando a productlist
